Rename `bt.params.al21c` -> `bt.params.cetbv2_amsr2`
