+++ pm_icecon/tests/unit/test_compute_bt_ic.py
@@ -1,7 +1,7 @@
 import datetime as dt
 
 from pm_icecon.bt.compute_bt_ic import _get_wx_params
-from pm_icecon.bt.params.a2l1c import (
+from pm_icecon.bt.params.cetbv2_amsr2 import (
     A2L1C_NORTH_PARAMS,
     _ret_parameters_amsru2_f_params,
 )
