+++ scripts/example_bt_script.py
@@ -11,7 +11,7 @@ from pm_tb_data.fetch.a2l1c_625 import get_a2l1c_625_tbs
 
 from pm_icecon._types import Hemisphere
 from pm_icecon.bt.compute_bt_ic import goddard_bootstrap
-from pm_icecon.bt.params.a2l1c import A2L1C_NORTH_PARAMS
+from pm_icecon.bt.params.cetbv2_amsr2 import A2L1C_NORTH_PARAMS
 from pm_icecon.config.models.bt import BootstrapParams
 
 if __name__ == '__main__':
